Update on audio ended implementation
