+++ .vscode/settings.json
@@ -1,26 +1,26 @@
 {
-    "eslint.validate": [
-        "javascript",
-        "javascriptreact",
-        "typescript",
-        "typescriptreact"
-    ],
-    "editor.codeActionsOnSave": {
-        "source.fixAll.eslint": true,
-        "source.fixAll.stylelint": true,
-    },
-    "editor.formatOnSave": true,
-    "typescript.format.enable": false,
-    "javascript.format.enable": false,
-    "cssVariables.lookupFiles": [
-        "src/styles/theme.scss",
-        "src/styles/themes/_light.scss", // include one of the theme file
-    ],
-    "prettier.disableLanguages": [
-        "javascript",
-        "javascriptreact",
-        "typescript",
-        "typescriptreact",
-        "scss"
-    ],
+  "eslint.validate": [
+    "javascript",
+    "javascriptreact",
+    "typescript",
+    "typescriptreact"
+  ],
+  "editor.codeActionsOnSave": {
+    "source.fixAll.eslint": true,
+    "source.fixAll.stylelint": true
+  },
+  // "editor.formatOnSave": true,
+  "typescript.format.enable": false,
+  "javascript.format.enable": false,
+  "cssVariables.lookupFiles": [
+    "src/styles/theme.scss",
+    "src/styles/themes/_light.scss" // include one of the theme file
+  ],
+  "prettier.disableLanguages": [
+    "javascript",
+    "javascriptreact",
+    "typescript",
+    "typescriptreact",
+    "scss"
+  ]
 }
